2 a-g: Final Versions
These are the final versions for Assignment 2 a-g. Use the SQL Code in
these files as reference for input in Sqlite3. result.txt was the file
I used to upload the proper answers to Coursera. I'm still not sure how
to do 2.6 (ie. 2f) and I need to do more research on that one.